admin/requirements.txt
flask==3.0.2
flask-login==0.6.3
werkzeug==3.0.1
pymongo==4.6.1
pandas==2.2.1
plotly==5.19.0
bcrypt==4.1.2
python-jose==3.3.0
python-dotenv==1.0.1